chore: replace legacy mock data
